LayoutTests/imported/w3c/web-platform-tests/html/browsers/offline/appcache/workers/appcache-worker.https-expected.txt

CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: null is not an object (evaluating 'worker.port.onmessage = r')
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: null is not an object (evaluating 'worker.port.onmessage = r')
CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: null is not an object (evaluating 'worker.port.onmessage = r')

PASS Dedicated worker of the cached script
PASS Dedicated worker of the fallbacked script
PASS Dedicated worker of the not-in-cache script
FAIL Shared worker of the cached script promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Shared worker of the fallbacked script promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
FAIL Shared worker of the not-in-cache script promise_test: Unhandled rejection with value: object "TypeError: null is not an object (evaluating 'worker.port.start')"
